server/app/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile '.\requirements.in'
#
aiosmtplib==3.0.2
    # via fastapi-mail
alt-profanity-check==1.5.2
    # via -r .\requirements.in
annotated-types==0.7.0
    # via pydantic
anonip==1.1.0
    # via -r .\requirements.in
anyio==4.7.0
    # via
    #   asyncer
    #   httpx
    #   starlette
    #   watchfiles
argon2-cffi==23.1.0
    # via -r .\requirements.in
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
asyncer==0.0.8
    # via fastapi-cdn-host
asyncmy==0.2.10
    # via -r .\requirements.in
asyncpg==0.30.0
    # via -r .\requirements.in
blinker==1.9.0
    # via fastapi-mail
certifi==2024.12.14
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via
    #   argon2-cffi-bindings
    #   cryptography
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via
    #   rich-toolkit
    #   typer
    #   uvicorn
colorama==0.4.6
    # via
    #   click
    #   uvicorn
cryptography==44.0.0
    # via
    #   -r .\requirements.in
    #   pyjwt
databases==0.9.0
    # via -r .\requirements.in
dnspython==2.7.0
    # via email-validator
email-validator==2.2.0
    # via
    #   fastapi
    #   fastapi-mail
fastapi[all]==0.115.6
    # via
    #   -r .\requirements.in
    #   fastapi-cache2
    #   fastapi-cdn-host
fastapi-cache2==0.2.2
    # via -r .\requirements.in
fastapi-cdn-host==0.8.3
    # via -r .\requirements.in
fastapi-cli[standard]==0.0.7
    # via fastapi
fastapi-mail==1.4.2
    # via -r .\requirements.in
greenlet==3.1.1
    # via sqlalchemy
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.7
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via
    #   fastapi
    #   fastapi-cdn-host
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
itsdangerous==2.2.0
    # via fastapi
jinja2==3.1.5
    # via
    #   fastapi
    #   fastapi-mail
joblib==1.4.2
    # via
    #   alt-profanity-check
    #   scikit-learn
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
numpy==2.2.1
    # via
    #   scikit-learn
    #   scipy
orjson==3.10.13
    # via fastapi
pendulum==3.0.0
    # via fastapi-cache2
pycparser==2.22
    # via cffi
pydantic==2.10.4
    # via
    #   fastapi
    #   fastapi-mail
    #   pydantic-extra-types
    #   pydantic-settings
pydantic-core==2.27.2
    # via pydantic
pydantic-extra-types==2.10.1
    # via fastapi
pydantic-settings==2.7.0
    # via
    #   fastapi
    #   fastapi-mail
pygments==2.18.0
    # via rich
pyjwt[crypto]==2.10.1
    # via -r .\requirements.in
pymysql==1.1.1
    # via -r .\requirements.in
pyotp==2.9.0
    # via -r .\requirements.in
python-dateutil==2.9.0.post0
    # via
    #   -r .\requirements.in
    #   pendulum
    #   time-machine
python-dotenv==1.0.1
    # via
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.20
    # via fastapi
pytz==2024.2
    # via -r .\requirements.in
pyyaml==6.0.2
    # via
    #   -r .\requirements.in
    #   fastapi
    #   uvicorn
redis==5.2.1
    # via -r .\requirements.in
requests==2.32.3
    # via -r .\requirements.in
rich==13.9.4
    # via
    #   -r .\requirements.in
    #   rich-toolkit
    #   typer
rich-toolkit==0.12.0
    # via fastapi-cli
scikit-learn==1.5.2
    # via alt-profanity-check
scipy==1.14.1
    # via scikit-learn
shellingham==1.5.4
    # via typer
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.36
    # via
    #   -r .\requirements.in
    #   databases
starlette==0.41.3
    # via
    #   fastapi
    #   fastapi-mail
threadpoolctl==3.5.0
    # via scikit-learn
time-machine==2.16.0
    # via pendulum
typer==0.15.1
    # via fastapi-cli
typing-extensions==4.12.2
    # via
    #   anyio
    #   fastapi
    #   fastapi-cache2
    #   pydantic
    #   pydantic-core
    #   pydantic-extra-types
    #   rich-toolkit
    #   sqlalchemy
    #   typer
tzdata==2024.2
    # via pendulum
ua-parser==1.0.0
    # via
    #   -r .\requirements.in
    #   user-agents
ua-parser-builtins==0.18.0.post1
    # via ua-parser
ujson==5.10.0
    # via fastapi
urllib3==2.3.0
    # via requests
user-agents==2.2.0
    # via -r .\requirements.in
uvicorn[standard]==0.34.0
    # via
    #   -r .\requirements.in
    #   fastapi
    #   fastapi-cache2
    #   fastapi-cli
watchfiles==1.0.3
    # via uvicorn
websockets==14.1
    # via uvicorn
